move upload vote batch api call into its own folder
